requirements.txt
Django==3.2
django-phonenumber-field==5.1.0
phonenumbers==8.12.21
python-dotenv==0.17.0
